provide a complete docker-based (manual) test environment.
Usage:

Prereq: docker on mac installed and the docker-machine running.

run tests/manual_run.py, which will bring up a docker container running
gkeepd and a mock mail server.  It will also create a docker container
with a prof account where you can run gkeep commands.  ssh keys are set up.

To get to the client, run command_line/sshclient from the command line.
You're on your own after that, because the prof account is empty other
than a helper script to upload the keys, the keys, and the git-keeper
cfg file.

The mock email server writes each email to a separate file with the recipient
and a sequential number as the filename.  The server folder is a docker
volume accessible locally in tests/tmp_mail...

If things don't work, try TestingEnvironment(debug_output=False), and look
at the output produced.

Notes:
- This is ready for automated testing, but I haven't tested that it works
that way.  My thinking is that we can add custom files to the server and/or
client to prep the system (this should be faster than running all the gkeep
commands).
- One challenge is that only keeper@gkserver is a docker mount.  When I
tried to mount at /home, the permissions weren't right and new users
could not log in.  This will limit our ability to check that user files are
created during testing...
- The permissions on tests/client_files/ssh/* and tests/server_files/ssh/*
must be 600 on the local machine.  I'm not sure git will respect those on
commit.