re-work Dockerfiles to avoid potential cache error.
This combines the apt-get update and apt-get install commands into a
single RUN statement.  Docker caches results, and so the apt-get update
is not executed simple because we add a new apt-get install.  When the
library is updated, docker won't see it.  By combining all apt-get actions
into one docke RUN statement, this issue is resolved - albeit with a slower
build time because all packages are re-installed if we add/remove anything.