implement submission testing.
This implementation has fully implements running tests directly on the
server.  Partial code is already present for running the tests within a
docker sandbox container.

If any error occurs while the tests are running the server will send an
announcement email to the student as well as a diagnostic email to the
faculty member associated with the course.