Improve/fetchinfo (#59)
* Refresh info after changes

Previously, when the client would query the server or fetch submissions it would
ask the server to generate a JSON file containing information about classes and
assignments. When the number of assignments in a class grows this slows down
fetching and querying considerably. Now, a new JSON file is created every time
something changes on the server, and the client just grabs the contents of the
most recent file. Only the 10 most recent files are kept for each faculty member.