Run assignment tests as a separate user
By default the tester user is named 'tester'. This is configurable
in server.cfg via the tester_user parameter of [gkeepd].

The tester user is created automatically if it does not exist.

run_action.sh is now created in the tester user's home directory.

The keeper user must now be able to run commands as any user
(not just root) without a password using sudo.