Speed up re-fetching assignments
Attempting to re-fetch an assignment was slow because the local git
hashes needed to be aquired for each repository and each student's
home directory on the server was aquired one-by-one. Now the local
git repository hashes are cached, and the students' home directories
are fetched in bulk from the server.