Use repo path from server info when fetching
Previously, git clone URLs were built up using the student's home
directory on the server and the names of the faculty, the class, and
the assignment. Since the path to the repo is already in the info
that is fetched from the server, there is no need to fetch the
students' home directories.