Handle command output that is not valid utf-8
gkeepcore.shell_command.run_command() uses decode() to convert the
output of shell commands from byte sequences to utf-8 before
returning. This used to raise a UnicodeDecodeError if the output
contains byte sequences that cannot be converted to utf-8. Now
the 'replace' error handler is used with decode() so that invalid
sequences are replaced with the � character.