Associate commit hashes with submissions
Previously, tests were always run on the latest commit of the master
branch of the student's repository. If a student pushed twice before
the server had a chance to process the first submission, tests would
be run twice on the code from the second submission.

Now the post-receive hook logs the hash of the latest commit, and
that hash is used to checkout that specific commit before running
tests.

This also means that tests will be run on a submission that is pushed
to any branch in the repository.