Improve performance when updating JSON information
Previously, every time the state of a faculty member's classes and
assignments changed on the server, gkeep performed a full scan of
that faculty member's classes and assignments, including collecting
the commit hashes of all of the students' submissions. On servers with
lots of classes and assignments, this started to take a long time
which caused a delay between the time that the state of the server was
updated and the time that the client could be aware of that change.

Now only the relevent information is updated when there is a state
change.