Reset server and client before every test and after all tests
Previously, individual commands were used to remove users, delete
files, stop the server, and delete email.  Each call had the overhead
of an ssh connection, and if a step was forgotten, it could invalidate
a subsequent test.

This version adds `reset_client.py` and `reset_server.py` that
ensure that both VMs are ready for the next test:

* (both VMs) Delete all users except `keeper`, `ubuntu`, and `vagrant`
* (gkserver) Stop `gkeepd` - if it is running
* (gkserver) Delete `gkeepd.log`, `snapshot.json`, `faculty.csv`, and
  `server.cfg` - if they exist.
* (gkserver) Delete any email messages saved in `/email`

These processes are called during the `Test Setup` for all tests and
then during the `Suite Teardown` - i.e. after all tests are executed.
This ensures that each test begins with the same VMs - and that both
VMs are "clean" when robot framework exits.