Add support for empty or ommitted CSV files with gkeep add
This commit allows the user to create a class before there are
any students.  This will be useful for faculty who work well
in advance of the semester.

The faculty member can either use an empty CSV file or run `gkeep add`
with just the name of the class (i.e. without the CSV file specified).

NOTE: Previously `gkeep add cs1 empty.csv` with an empty csv would
succeed - which caused confusion when my intern tried to test
that it would fail.