Improve class modification
Students can now be cleanly re-added to a class. When a student is
removed from a class, they are marked as deactivated and will not
have repositories set up on their behalf for future assignments.
If a student is re-added to a class, they are re-activated and any
assignments that were published since deactivation will be set
up for the student.

Rather than uploading the entire roster as a CSV to the server to modify
a class, the client now detects and devides the modification task up into
3 phases, and sends requests for each phase separately:
- Adding students
- Removing students
- Renaming students

Adding a new class also now happens in 2 phases, first the new class is
created, then students are added to the class.

The database was re-organized so that the first and last names of a
student user are stored separately for each class. Some other
re-factoring of the database code happened as well, along with changes
to the database tests.