Add support for diabling assignments
When disabling an assignment, students will be sent an email notifying
them that the assignment is disabled. Attempts to push submissions for
the assignment will result in a similar email, and no tests will be run.

The assignment will show up in queries but will be marked as disabled.

Faculty cannot fetch, trigger, update, delete, or publish disabled
assignments.