Reduce the priority of bulk emails in the email queue
Adding a new class roster and publishing an assignment both cause
a batch of emails to be added to the sending queue. Since email
sending is rate limited, other emails will be delayed. This patch
lowers the priority of new account emails and new assignment emails,
but only for students. This increases the responsiveness for actions
where a single email is expected right away, such as test results
and new assignment emails that go to faculty upon uploading.