Fix gkeepd's top level exception handler
gkeepd's main loop waits for events and calls upon handlers to handle
the events. A properly written handler should catch any exceptions that
are raised while handling an event, but in case an exception is not
caught at the proper time it will be caught in gkeepd's main loop so
that the program does not crash.

There was an error in the way that tracebacks were extracted and
reported when these exceptions occur, which is fixed here.