Allow sending results as HTML with the pre tag
An HTML version of submission result emails can now be sent which
escapes results using html.escape() and wraps the content in
<pre></pre> tags. This indicates to email clients that the results
should be shown as a code block using a monospace font.

This is disabled by default, and can be enabled by setting use_html
to true in the email section of server.cfg.