Adjust git usage due to new git behavior
Due to a vulnerability, git no longer allows commands to be issued
when the user issuing the commands is not the same as the owner of
the repositories involved. This required changes to the way that
gkeepd pushes to reports repositories, and the way that it retrieves
hashes from repositories.