Move event handling to a new thread
Previously when gkeepd was shut down, some log events could be lost if
they were still in the event handler queue, because the loop that
processed items in the queue immediately exited on shutdown. How the
handling of these events is moved to its own thread which is shut down
in a similar manner to the other threads. All events will be handled
before gkeepd shuts down.