Change email matching and username validation
Checks for duplicate email addresses in the database are now done in a
case-insensitive manner, though the cases of the characters in the originally
added email address are preserved in case the email server handles the username
portion of the email address in a case-sensitive manner.

Additionally, it used to be the case that a student email address was rejected if
its username portion was not also a valid Linux username. This is no longer the
case, since the string is converted to an appropriate Linux username when creating
the user.