Add newlines to mysmtpd.py email bodies
The most straightforward way to read emails on gkserver is with cat,
and since some emails do not contain a newline at the end of the email
body it messes with the prompt and makes the email hard to read. For
HTML emails, both the plain text and HTML portions were not delineated
with newlines either. Now there will always be exactly 2 newlines after
the body of each email, both for plain text and HTML portions.