Update new user naming policy
New users will now always get a new account, even if an account exists with
the same username as their email username and the account is not used by
an existing git-keeper user. In the case of username clashes, the new
user's username is modified by appending incrementing positive integers
to the email username until a unique name is found.

Additionally, users may now be added whose email usernames are keeper,
tester, faculty, or student. Dummy accounts are added to the database
to reserve these names.