Blacklist the assignment name tests
Since the student's submission repository is cloned into the temporary
directory that contains the tests folder, testing will not work properly
if there is an assignment named 'tests', so 'tests' is no longer a valid
name for an assignment.

Tests were added to ensure upload fails if an assignment is named 'tests'
or 'all'.