Update test environment and stop quoting robot strings
The testing Vagrant environments for both client and server were
updated to use ubuntu/jammy64. This also required a new approach
for checking if an assignment's tests are being run within a
Docker container. It now checks for the existance of /.dockerenv

The Python code to execute scripts on the VMs did not escape
characters in command line arguments. This meant that all strings
containing spaces needed to be quoted before passed to the methods
that execute the scripts. Now all such strings are escaped using
shlex.quote().