Adding internal support for HTML and Markdown emails
This does not effect any current usage of Emails and the default behavior is to use plain text just as before. The new system allows for supplying a pre-setup MIMEBase object to use as the body of the email instead of the literal text. This setup provides functions for generating the bodies from plain text, HTML, or Markdown.