Fix issue caused by Git update
By default, git no longer allows a user to clone or push to a
repository owned by a different user, even if file permissions are
set up to allow it. This breaks git-keeper functionality. The
workaround is to either mark individual directories as safe in a
git configuration file, or mark all directories as safe. We have
chosen the latter, because we believe that the permissions on a
git-keeper server prevent the insecure scenario that led to the change
in the first place.

Now gkeepd will check if the system-wide Git configuration marks all
directories as safe on startup, and if that is not the case, it adds
changes the configuration to mark all directories as safe using the
following command, run with sudo:

    git config --system --set safe.directory '*'

The "Valid Setup" test in gkeepd_launch.robot now checks that the
configuration option is set appropriately after confirming that gkeepd
is running. When the tests reset the server, the file /etc/gitconfig
is removed, forcing it to be created again when gkeepd is launched.

A note was added to the admin user documentation about the git
configuration.